fix: ajustes no back para gravação dos pontos de parada

+++ Repositories/Interface/IStopPointRepository.cs
@@ -5,8 +5,8 @@ namespace ESProjeto_Back.Repositories.Interface
 {
     public interface IStopPointRepository
     {
-        public GetCloseStopPointsDto GetPointsOnRadiousOf(float latitude, float longitude, float radius);
+        public GetCloseStopPointsDto GetPointsOnRadiousOf(double latitude, double longitude, float radius);
         public StopPoint? GetStopPointInfo(Guid id);
-        public Guid SetNewStopPoint(StopPoint stopPoint, float latitude, float longitude);
+        public Guid SetNewStopPoint(StopPoint stopPoint, double latitude, double longitude);
     }
 }
